pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tec-tgcr"
version = "0.1.0"
description = "The Elidoras Codex agent platform"
readme = "README.md"
authors = [
    { name = "TEC-The-Elidoras-Codex" }
]
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.6",
    "rich>=13.7",
    "typer>=0.12",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "python-docx>=1.1",
]

[project.scripts]
tec-agent = "tec_tgcr.cli:app"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = [
    "tests",
]
